Both Pages Work Perfectly(One Problem)
Having problem with the validation, seems to all work, but comes up with
error page if value is empty.
